[HomeWizard] Process review comments (openhab#9825)
Processed the comments from the 3rd review.
