+++ bundles/org.openhab.binding.homewizard/src/main/resources/OH-INF/binding/binding.xml
@@ -4,6 +4,7 @@
 	xsi:schemaLocation="https://openhab.org/schemas/binding/v1.0.0 https://openhab.org/schemas/binding-1.0.0.xsd">
 
 	<name>HomeWizard Binding</name>
-	<description>This is the binding for HomeWizard.</description>
+	<description>This binding provides access to the DSMR data provided by the HomeWizard Wi-Fi P1 meter on it's local HTTP
+		interface.</description>
 
 </binding:binding>
